+++ validation-tests/InstanceTests.cs
@@ -1,11 +1,14 @@
 ﻿using System;
+using CSGOStats.Extensions.Validation.Tests.Model;
 using FluentAssertions;
 using Xunit;
 
-namespace CSGOStats.Infrastructure.Validation.Tests
+namespace CSGOStats.Extensions.Validation.Tests
 {
     public class InstanceTests
     {
+        private static readonly Random Random = new Random();
+
         [Fact]
         public void NotNullForClassesTest()
         {
@@ -38,24 +41,52 @@ namespace CSGOStats.Infrastructure.Validation.Tests
             }
         }
 
+        [Fact]
+        public void UnwantedValuePositiveTest()
+        {
+            {
+                var instance = Random.Next();
+                var comparable = Random.Next();
+                instance.AnythingBut(comparable, nameof(instance)).Should().Be(instance);
+            }
+
+            {
+                var instance = GetDefaultClassStatefulInstance();
+                var comparable = GetDefaultClassStatefulInstance();
+                instance.AnythingBut(comparable, nameof(instance)).Data.Should().Be(instance.Data);
+            }
+        }
+
+        [Fact]
+        public void UnwantedValueNegativeTest()
+        {
+            {
+                var instance = Random.Next();
+                var comparable = instance;
+                const string instanceName = nameof(instance);
+                var exception = Record.Exception(() => instance.AnythingBut(comparable, instanceName)).Should().BeOfType<UnwantedState>().Subject;
+                exception.InstanceName.Should().Be(instanceName);
+                exception.Value.Should().BeOfType<int>().Which.Should().Be(instance);
+            }
+
+            {
+                var instance = GetDefaultClassStatefulInstance();
+                var comparable = new TestClass(instance.Data);
+                const string instanceName = nameof(instance);
+                var exception = Record.Exception(() => instance.AnythingBut(comparable, instanceName)).Should().BeOfType<UnwantedState>().Subject;
+                exception.InstanceName.Should().Be(instanceName);
+                exception.Value.Should().BeOfType<TestClass>().Which.Data.Should().Be(instance.Data);
+            }
+        }
+
         private static TestClass GetDefaultClassInstance() => new TestClass();
+
+        private static TestClass GetDefaultClassStatefulInstance() => new TestClass(Random.Next());
 
         private static TestClass GetNullClassInstance() => null;
 
         private static TestStruct? GetDefaultStructInstance() => new TestStruct();
 
         private static TestStruct? GetNullStructInstance() => null;
-    }
-
-    public interface ITestInterface
-    {
-    }
-
-    public class TestClass : ITestInterface
-    {
-    }
-
-    public struct TestStruct : ITestInterface
-    {
     }
 }