+++ validation-tests/Model/ITestInterface.cs
@@ -1,0 +1,6 @@
+﻿namespace CSGOStats.Extensions.Validation.Tests.Model
+{
+    public interface ITestInterface
+    {
+    }
+}